+++ documentation/source/conf.py
@@ -95,6 +95,7 @@ extensions = [
     "sphinx.ext.intersphinx",
     "sphinx.ext.viewcode",
     "sphinx_autodoc_typehints",
+    "inherited_typehints",
     "sphinxcontrib.bibtex",
 ]